ml_worker/requirements.txt
celery[redis]
wheel
certifi
eventlet
flask
flask-restx
flask_cors
flask-talisman
itsdangerous
numpy
pandas
requests
webcolors
werkzeug
gunicorn
boto3
openpyxl
xlsxwriter
textract-trp
gevent
opencv-contrib-python
opencv-python
opencv-python-headless
python-dotenv
python-Levenshtein
thefuzz
PyPDF2
PyMUpdf
pillow
pytesseract
scikit-image
scipy
transformers
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
google-cloud-vision==3.4.4
